// src/models/userData.ts
export type userData = {
    user_id: number;
    name: string;
    last_name?: string;
    date_of_birth?: string;
    user_type?: string;
    email: string;
    password: string;
    active: boolean;
    start_date: string;
    end_date?: string;
    created_by?: string;
    created_at: string;
    updated_at?: string;
    cost_model_id?: number;
    user_type_id?: number;
}
  